odds_analysis_system/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
gunicorn==21.2.0
requests==2.31.0
plotly==5.18.0
dash==2.14.0
dash-bootstrap-components==1.5.0
scikit-learn==1.3.0
scipy==1.11.3
APScheduler==3.10.4
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Caching==2.1.0
Flask-Session==0.5.0
python-dateutil==2.8.2
Flask-Limiter==3.5.0
Flask-Cors==4.0.0